+++ Runtime/Scripts/Variable Handlers/VariableHandlerList.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SLIDDES.Modular
+{
+    [AddComponentMenu("SLIDDES/Modular/Variable Handlers/Variable Handler List")]
+    public class VariableHandlerList : VariableHandler<List<ScriptableObject>>
+    {
+        
+    }
+}