+++ Runtime/Scripts/Variable Handlers/VariableHandlerVector2.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SLIDDES.Modular
+{
+    [AddComponentMenu("SLIDDES/Modular/Variable Handlers/Variable Handler Vector2")]
+    public class VariableHandlerVector2 : VariableHandler<Vector2>
+    {
+        
+    }
+}