+++ Runtime/Scripts/Variable Handlers/VariableHandlerVector3.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SLIDDES.Modular
+{
+    [AddComponentMenu("SLIDDES/Modular/Variable Handlers/Variable Handler Vector3")]
+    public class VariableHandlerVector3 : VariableHandler<Vector3>
+    {
+        
+    }
+}