+++ Runtime/Scripts/Variable Handlers/VariableHandlerVector3Int.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SLIDDES.Modular
+{
+    [AddComponentMenu("SLIDDES/Modular/Variable Handlers/Variable Handler Vector3Int")]
+    public class VariableHandlerVector3Int : VariableHandler<Vector3Int>
+    {
+        
+    }
+}